+++ CHANGES.rst
@@ -111,8 +111,10 @@ Changelog
 - Fixed `meeting.committees` default value, ignore MeetingConfig.committees
   that use `enabled="item_only"`.
   [gbastien]
-- Invalidate meeting `actionspanel` caching when meeting contains or
-  no more contains items so the `Delete` action is displayed or not.
+- Finally fixed invalidating meeting `actionspanel` caching when meeting
+  contains/does not contain items so the `Delete` action is handled correctly.
+  `Meeting.number_of_items` replaced parameter `as_int=False` by `as_str=False`
+  as we only want it to be str for JS.
   [gbastien]
 
 4.2rc34 (2022-09-29)